refactor code to arraynge structured tree
