+++ getNamesAsync/getLogos.js
